+++ fr.centralesupelec.edf.riseclipse.iec61850.scl/src/fr/centralesupelec/edf/riseclipse/iec61850/scl/VoltageLevel.java
@@ -108,12 +108,12 @@ public interface VoltageLevel extends EquipmentContainer {
      * @return the value of the '<em>Num Phases</em>' attribute.
      * @see #isSetNumPhases()
      * @see #unsetNumPhases()
-     * @see #setNumPhases(Byte)
+     * @see #setNumPhases(Integer)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.scl.SclPackage#getVoltageLevel_NumPhases()
      * @model unsettable="true"
      * @generated
      */
-    Byte getNumPhases();
+    Integer getNumPhases();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.scl.VoltageLevel#getNumPhases <em>Num Phases</em>}' attribute.
@@ -125,7 +125,7 @@ public interface VoltageLevel extends EquipmentContainer {
      * @see #getNumPhases()
      * @generated
      */
-    void setNumPhases( Byte value );
+    void setNumPhases(Integer value);
 
     /**
      * Unsets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.scl.VoltageLevel#getNumPhases <em>Num Phases</em>}' attribute.
@@ -133,7 +133,7 @@ public interface VoltageLevel extends EquipmentContainer {
     * <!-- end-user-doc -->
      * @see #isSetNumPhases()
      * @see #getNumPhases()
-     * @see #setNumPhases(Byte)
+     * @see #setNumPhases(Integer)
      * @generated
      */
     void unsetNumPhases();
@@ -145,7 +145,7 @@ public interface VoltageLevel extends EquipmentContainer {
      * @return whether the value of the '<em>Num Phases</em>' attribute is set.
      * @see #unsetNumPhases()
      * @see #getNumPhases()
-     * @see #setNumPhases(Byte)
+     * @see #setNumPhases(Integer)
      * @generated
      */
     boolean isSetNumPhases();
