+++ fr.centralesupelec.edf.riseclipse.iec61850.scl/src/fr/centralesupelec/edf/riseclipse/iec61850/scl/impl/LineImpl.java
@@ -163,7 +163,7 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
      * @generated
      * @ordered
      */
-    protected static final Byte NUM_PHASES_EDEFAULT = null;
+    protected static final Integer NUM_PHASES_EDEFAULT = null;
 
     /**
      * The cached value of the '{@link #getNumPhases() <em>Num Phases</em>}' attribute.
@@ -173,7 +173,7 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
      * @generated
      * @ordered
      */
-    protected Byte numPhases = NUM_PHASES_EDEFAULT;
+    protected Integer numPhases = NUM_PHASES_EDEFAULT;
 
     /**
      * This is true if the Num Phases attribute has been set.
@@ -538,7 +538,7 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
      * <!-- end-user-doc -->
      * @generated
      */
-    public Byte getNumPhases() {
+    public Integer getNumPhases() {
         return numPhases;
     }
 
@@ -547,8 +547,8 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
      * <!-- end-user-doc -->
      * @generated
      */
-    public void setNumPhases(Byte newNumPhases) {
-        Byte oldNumPhases = numPhases;
+    public void setNumPhases(Integer newNumPhases) {
+        Integer oldNumPhases = numPhases;
         numPhases = newNumPhases;
         boolean oldNumPhasesESet = numPhasesESet;
         numPhasesESet = true;
@@ -562,7 +562,7 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
      * @generated
      */
     public void unsetNumPhases() {
-        Byte oldNumPhases = numPhases;
+        Integer oldNumPhases = numPhases;
         boolean oldNumPhasesESet = numPhasesESet;
         numPhases = NUM_PHASES_EDEFAULT;
         numPhasesESet = false;
@@ -707,7 +707,7 @@ public class LineImpl extends GeneralEquipmentContainerImpl implements Line {
                 setNomFreq((BigDecimal)newValue);
                 return;
             case SclPackage.LINE__NUM_PHASES:
-                setNumPhases((Byte)newValue);
+                setNumPhases((Integer)newValue);
                 return;
         }
         super.eSet(featureID, newValue);
